fix(server): break schedule enum circular imports
Move schedule enums into a dependency-free module and update schedule and course journal imports so unit tests can load schemas without circular initialization failures.
