+++ server/src/courses/journal/dto/create-lesson-journal-entry.dto.ts
@@ -15,7 +15,7 @@ import {
   Min,
   ValidateNested,
 } from 'class-validator';
-import { ScheduleEntryType } from '../../../schedule/schemas';
+import { ScheduleEntryType } from '../../../schedule/schedule.enums';
 import { AttendanceStatus } from '../../schemas';
 
 export class LessonAttendanceDto {
