+++ server/src/schedule/schemas/index.ts
@@ -1,2 +1,3 @@
+export * from '../schedule.enums';
 export * from './schedule-entry.schema';
 export * from './schedule-template.schema';
